Move function to file
