+++ src/main/kotlin/com/daveme/chocolateCakePHP/controller/ControllerFieldTypeProvider.kt
@@ -1,7 +1,7 @@
 package com.daveme.chocolateCakePHP.controller
 
 import com.daveme.chocolateCakePHP.*
-import com.daveme.chocolateCakePHP.cake.CakeModel.isCakeTwoModelClass
+import com.daveme.chocolateCakePHP.cake.isCakeTwoModelClass
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiElement
 import com.jetbrains.php.PhpIndex
